refactor: refrences forms and components Link T-12340
